Add frequency table
